requirements.txt
alembic==1.13.1
blinker==1.8.2
click==8.1.7
exceptiongroup==1.2.1
Flask==3.0.3
Flask-Admin==1.6.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
git-filter-repo==2.38.0
greenlet==3.0.3
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.1
pluggy==1.5.0
PyMySQL==1.0.2
pytest==8.2.2
python-dotenv==1.0.1
SQLAlchemy==2.0.30
tomli==2.0.1
typing_extensions==4.12.2
Werkzeug==3.0.3
WTForms==3.1.2
